fix: rerendering issues in sample app
