+++ examples/SampleApp/App.tsx
@@ -75,7 +75,7 @@ const Drawer = createDrawerNavigator();
 const Stack = createStackNavigator<StackNavigatorParamList>();
 const UserSelectorStack = createStackNavigator<UserSelectorParamList>();
 const App = () => {
-  const { chatClient, isConnecting, loginUser, logout, switchUser, unreadCount } = useChatClient();
+  const { chatClient, isConnecting, loginUser, logout, switchUser } = useChatClient();
   const colorScheme = useColorScheme();
   const streamChatTheme = useStreamChatTheme();
 
@@ -141,7 +141,7 @@ const App = () => {
             dark: colorScheme === 'dark',
           }}
         >
-          <AppContext.Provider value={{ chatClient, loginUser, logout, switchUser, unreadCount }}>
+          <AppContext.Provider value={{ chatClient, loginUser, logout, switchUser }}>
             {isConnecting && !chatClient ? (
               <LoadingScreen />
             ) : chatClient ? (
